[REF] Fix code following review on PR
