+++ project_assignment/README.rst
@@ -1,0 +1,13 @@
+Project Assignment
+=================
+
+Automatically reassign task to specified partner depending on stage
+-------------------------------------------------------------------
+    * For each stage, the partner can be specified in stage,
+        then in project and finally task itself
+    * We use partner instead of user for more flexibility
+    * Use base inherit config
+
+## Dependencies ##
+- base_recursive_model https://github.com/OCA/server-tools
+- project
