+++ project_assignment/data/project_task_type_demo.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+
+        <!-- Types -->
+
+        <record id="project.project_tt_analysis" model="project.task.type">
+            <field name="partner_id" ref="base.partner_root"/>
+        </record>
+        <record id="project.project_tt_specification" model="project.task.type">
+            <field name="partner_id" ref="base.partner_root"/>
+        </record>
+        <record id="project.project_tt_development" model="project.task.type">
+            <field name="partner_id" ref="base.partner_demo"/>
+        </record>
+
+    </data>
+</openerp>
