+++ project_assignment/project_project_view.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+
+        <!-- Project -->
+
+        <record id="view_project_project_tree_assignpartner" model="ir.ui.view">
+            <field name="name">project.project.tree.assignpartner</field>
+            <field name="model">project.project</field>
+            <field name="inherit_id" ref="project.view_project"/>
+            <field name="arch" type="xml">
+                <field name="user_id" position="replace">
+                    <field name="manager_partner_id"/>
+                    <field name="user_id" invisible="1"/>
+                </field>
+            </field>
+        </record>
+
+        <record id="view_project_project_form_assignpartner" model="ir.ui.view">
+            <field name="name">project.project.form.assignpartner</field>
+            <field name="model">project.project</field>
+            <field name="inherit_id" ref="project.edit_project"/>
+            <field name="arch" type="xml">
+                <field name="user_id" position="replace">
+                    <field name="manager_partner_id"/>
+                    <field name="user_id" invisible="1"/>
+                </field>
+                <field name="type_ids" position="after">
+                    <group>
+                        <field name="assigned_partner_config_ids" context="{'default_model': 'project.project'}"/>
+                        <field name="assigned_partner_config_result_ids"/>
+                    </group>
+                </field>
+            </field>
+        </record>
+
+    </data>
+</openerp>
