+++ project_assignment/project_task_type_view.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+
+        <!-- Stage -->
+
+        <record id="view_project_task_type_form_assignpartner" model="ir.ui.view">
+            <field name="name">project.task.type.form.assignpartner</field>
+            <field name="model">project.task.type</field>
+            <field name="inherit_id" ref="project.task_type_edit"/>
+            <field name="arch" type="xml">
+                <field name="sequence" position="after">
+                    <field name="partner_id"/>
+                </field>
+            </field>
+        </record>
+
+    </data>
+</openerp>
